refactor: 移除 TabDisplay 和 TexEditor 组件，优化代码结构
